+++ semgrep/tests/conftest.py
@@ -67,7 +67,7 @@ def _run_semgrep(
     :param config: what to pass as --config's value
     :param target_name: which directory within ./e2e/targets/ to scan
     :param options: additional CLI flags to add
-    :param output_format: which format to use, valid options are normal, json, and sarif
+    :param output_format: which format to use, valid options are normal, json, junit_xml and sarif
     :param stderr: whether to merge stderr into the returned string
     """
     if options is None:
@@ -83,6 +83,8 @@ def _run_semgrep(
 
     if output_format == "json":
         options.append("--json")
+    elif output_format == "junit-xml":
+        options.append("--junit-xml")
     elif output_format == "sarif":
         options.append("--sarif")